Change var type from float to double

+++ CalculatorFunctions/add.c
@@ -2,17 +2,17 @@
 
 int main() {
 
-    float firstNumber, secondNumber, sum;
+    double firstNumber, secondNumber, sum;
 
     printf("First number: ");
-    scanf("%f", &firstNumber);
+    scanf("%lf", &firstNumber);
     printf("Second number: ");
-    scanf("%f", &secondNumber);
+    scanf("%lf", &secondNumber);
 
     sum = firstNumber + secondNumber;
 
-    printf("\n%.2f + %.2f = %.2f\n", firstNumber, secondNumber, sum);
-    printf("The sum is: %.2f\n", sum);
+    printf("\n%.2lf + %.2lf = %.2lf\n", firstNumber, secondNumber, sum);
+    printf("The sum is: %.2lf\n", sum);
 
     return 0;
 }
